UIU Bus
It's a beginner level javaFX project. Here has a bus with multiple functionalities. It can move with the instructions of buttons. "Left" button is to move to left, "Right" is to move right and so on. The "Speed" button is for speeding up the movement of the bus, "Slow" is for slowing down the movement, and "Normal" is for resetting the speed to default. I will attach the fxml file and HelloApplication file to run the code.